requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langgraph
langchain-google-genai
langchain-community
sentence-transformers
chromadb
ollama
pydantic
langchain-huggingface